requirements.txt
pandas
sqlalchemy
python-telegram-bot
requests
python-dotenv
